+++ examples/Domain/Course.php
@@ -25,12 +25,21 @@ class Course
      */
     private $language;
 
-    public function __construct(string $name, string $description, int $credit, string $language)
+    /**
+     * @var SchoolClass[]
+     */
+    private $classes;
+
+    /**
+     * @param SchoolClass[] $classes
+     */
+    public function __construct(string $name, string $description, int $credit, string $language, array $classes)
     {
         $this->name = $name;
         $this->description = $description;
         $this->credit = $credit;
         $this->language = $language;
+        $this->classes = $classes;
     }
 
     public function getName()
@@ -51,5 +60,13 @@ class Course
     public function getLanguage(): string
     {
         return $this->language;
+    }
+
+    /**
+     * @return SchoolClass[]
+     */
+    public function getClasses(): array
+    {
+        return $this->classes;
     }
 }