+++ examples/Infrastructure/Factory/StudentFactory.php
@@ -1,0 +1,55 @@
+<?php
+declare(strict_types=1);
+
+namespace WoohooLabs\Worm\Examples\Infrastructure\Factory;
+
+use DateTimeImmutable;
+use WoohooLabs\Worm\Examples\Domain\Name;
+use WoohooLabs\Worm\Examples\Domain\Student;
+use WoohooLabs\Worm\Examples\Infrastructure\Model\StudentModel;
+use WoohooLabs\Worm\Execution\IdentityMap;
+
+class StudentFactory extends AbstractFactory
+{
+    /**
+     * @var StudentModel
+     */
+    private $model;
+
+    public function __construct(IdentityMap $identityMap, StudentModel $model)
+    {
+        parent::__construct($identityMap);
+        $this->model = $model;
+    }
+
+    /**
+     * @return Student[]
+     */
+    public function createStudents(array $entities): array
+    {
+        return array_map(
+            function (array $entity) {
+                return $this->createStudent($entity);
+            },
+            $entities
+        );
+    }
+
+    public function createStudent(array $entity): Student
+    {
+        $factory = function (array $entity): Student {
+            return new Student(
+                $entity[$this->model->id],
+                new Name(
+                    $entity[$this->model->first_name],
+                    $entity[$this->model->last_name]
+                ),
+                $entity[$this->model->introduction],
+                $entity[$this->model->birthday] ? new DateTimeImmutable($entity[$this->model->birthday]) : null,
+                $entity[$this->model->gender] ?? null
+            );
+        };
+
+        return $this->createObject($this->model, $entity, $factory);
+    }
+}